add frontend telegram integration

+++ src/backend/messenger/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Integer, String
+from sqlalchemy import Boolean, BigInteger, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -10,5 +10,5 @@ class UserModel(BaseModel):
     id: Mapped[int] = mapped_column(primary_key=True)
     username: Mapped[str] = mapped_column(String(), unique=True)
     password_hash: Mapped[str] = mapped_column(String())
-    telegram_id: Mapped[int] = mapped_column(Integer(), nullable=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger(), nullable=True)
     active: Mapped[bool] = mapped_column(Boolean(), server_default="1")
